code generation wrappers refactoring and protocol test clients generation
