+++ tools/scripts/codegen/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0.
